backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hsinpaohuang <hsinpaohuang@gmail.com>"]
readme = "README.md"
packages = [{ include = "**/*" }]

[tool.taskipy.tasks]
dev = "cd /code/backend/backend; uvicorn main:app --host 0.0.0.0 --reload --reload-include /code/backend"
prod = "cd /code/backend/backend; uvicorn main:app"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
beanie = "^1.25.0"
aiohttp = "^3.9.3"
pydantic-settings = "^2.1.0"
fastapi-users = {extras = ["beanie", "oauth"], version = "^13.0.0"}
nltk = "^3.9.1"
numpy = "^2.1.1"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
